requirements.txt
django
django-rest-framework
dj_database_url